# utilities/openapi.yaml
openapi: 2.0.0
info:
  title: "Mozilla Doorman"
  description: |
    Doorman is an **authorization micro-service** that allows to checks if an arbitrary subject is allowed to perform an action on a resource, based on a set of rules (policies).

    ### Workflow with OpenID

    When the identity provider is configured for a service (*recommended*), a typical workflow is:

    1. Users obtain an access token from an Identity Provider (eg. Auth0)
    1. They use it to call the service API endpoint
    1. The service posts an authorization request on *Doorman* to check if the user is allowed
    1. *Doorman* uses the `Origin` request header to select the set of policies
    1. *Doorman* builds a list of strings (principals) for this user, and matches the policies

    *Doorman* will verify that the specified Access token (or ID token) is valid.

    The authorization request *principals* will be built from the user profile information:

    * `"sub"`: `userid:{}`
    * `"email"`: `email:{}` (*optional*)
    * `"groups"`: `group:{}, group:{}, ...` (*optional*)

    ### Without OpenID

    If the identity provider is not configured for a service, no authentication is required and the principals
    are posted in the authorization body.

    A typical workflow in this case would be:

    1. Users call the service API endpoint
    1. The service authenticates the user and builds the list of principals
...

    ### API Summary

...

    **Request**:

    ```HTTP

    POST /allowed HTTP/1.1
...

...
paths:
  /allowed:
    post:
      summary: Check authorization request
      description: |
        Are those `principals` allowed to perform this `action` on this `resource` in this `context`?

        With OpenID enabled, the principals are either read from the Identity Provider user info endpoint or directly from the JSON Web Token payload
        if an ID token is provided.

      operationId: "allowed"
      consumes:
        - application/json
      produces:
      - "application/json"
      parameters:
        - in: header
          name: Origin
          type: string
          description: |
            The service identifier (eg. `https://api.service.org`). It must match one of the known service from the policies files.

        - in: header
          name: Authorization
          type: string
          description: |
            With OpenID enabled, a valid Access token (or JSON Web ID Token) must be provided in the `Authorization` request header.
            (eg. `Bearer eyJ0eXAiOiJKV1QiLCJhbG...9USXpOalEzUXpV`)

        - in: body
          description: |
            Authorization request as JSON.

            Note that **every field is optional**.

          required: true
          schema:
            type: object
            properties:
              principals:
                description: |
                  **Only without OpenID**

                  Arbitrary list of strings (eg. `userid:alice`, `group:editors`).

                type: array
                items:
                  type: string
              action:
                description: Any domain specific action (eg. `read`, `delete`, `signoff`)
                type: string
              resource:
                description: Any resource (eg. `blocklist`, `/update/rules/35`)
                type: string
              context:
                description: |
                  The context can contain any extra information to be matched in policies conditions.
                  The context field `remoteIP` will be forced by the server.
                  The values provided in the `roles` context field will expand the principals with extra `role:{}` values.

                type: object
                properties:
                  roles:
                    type: array
                    items:
                      type: string
          example:
            principals: ["userid:ldap|ada", "email:ada@lau.co"]
            action: create
            resource: comment
            context:
              env:
                - stage
              roles:
                - changer
      responses:
        "400":
          description: "Missing headers or invalid posted data."
          schema:
            type: object
            properties:
              message:
                type: string
          example:
            message: Missing `Origin` request header
        "401":
          description: "OpenID token is invalid."
        "200":
          description: "Return whether it is allowed or not."
          schema:
            type: object
            properties:
              allowed:
                type: boolean
              principals:
                type: array
                items:
                  type: string
          example:
            allowed: true
            principals: ["userid:ldap|ada", "email:ada@lau.co", "tag:mayor", "role:changer"]
      tags:
      - Doorman
...